requirements.txt
numpy
opencv-python
imutils
Flask==1.1.2
imageio==2.9.0
tensorflow==2.5.0
keras
pyjwt
Pillow==8.2.0
sklearn
